examples/summarizer-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via starlette
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via
    #   requests
    #   wale-sdk
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
    #   wale-sdk
click==8.1.3
    # via uvicorn
fastapi==0.95.1
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
    #   wale-sdk
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.27.6
    # via -r requirements.in
pydantic==1.10.7
    # via
    #   -r requirements.in
    #   fastapi
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.30.0
    # via
    #   openai
    #   wale-sdk
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
tqdm==4.65.0
    # via openai
typing-extensions==4.5.0
    # via pydantic
urllib3==2.0.2
    # via
    #   requests
    #   wale-sdk
uvicorn==0.22.0
    # via -r requirements.in
wale-sdk==0.1.2
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
